% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas_Type.R
\name{CaSilico}
\alias{CaSilico}
\title{In silico crRNA design for different CRISPR-Cas subtypes}
\usage{
CaSilico(
  Results_folder_name = "CaSilico_output",
  fasta_file_location = NULL,
  Accession_Number = NULL,
  seq_cordinate = list(chromosome = NA, start = NA, end = NA, strand = NA, species =
    NA),
  CRISPR_Types,
  method = 1,
  mismatch_threshold = 0.95,
  off_Target = F,
  ask_question = T,
  Organism = NULL,
  local_file_for_off_target = F,
  location_of_off_target_sequene_file = c("path1", "path2"),
  name_of_off_target_sequence_file = c("Organism1", "Organism2")
)
}
\arguments{
\item{Results_folder_name}{final results folder name.}

\item{fasta_file_location}{path and name of fasta file or Accession_Number.}

\item{Accession_Number}{genome/gene/transcript identifier(s) (RefSeq or ENSEMBL ID) ex: "NM_000546".}

\item{seq_cordinate}{A list of genomic coordinate includes the chromosome, Start, End, Strand  and species. CaSilico retrieves the sequence of each of these coordinates.}

\item{CRISPR_Types}{A vector of CRSPR subtypes :c("casVI_A","casVI_B","casVI_D","casV_A","casV_B","casV_F1").}

\item{method}{two different methods (1 or 2) for detection of conserved regions.}

\item{off_Target}{A Boolean TRUE/FALSE variable to analysis off-Target for desired organisms by default FALSE.}

\item{ask_question}{A Boolean TRUE/FALSE variable to answer a number of questions during off target prediction.}

\item{Organism}{A vector of desired organisms for off target prediction like : c("Homo_sapiens","bos_tarus",...).}

\item{local_file_for_off_target}{A Boolean TRUE/FALSE variable to analysis off-Target by local fasta file by default FALSE.}

\item{location_of_off_target_sequene_file}{A vector of paths for local off target prediction ex:c("/yourpath1/filename1","/yourpath2/filename2",...).}

\item{name_of_off_target_sequence_file}{A vector including desired names related to the "location_of_off_target_sequene_file" Casilico puts sequence1, sequence2, ... by default if the user doesn't define names.}

\item{conservation_threshold}{described threshold (between 0 and 1) in detection of conserved regions by default:0.95.}
}
\value{
plots and html files of results .
}
\description{
In silico crRNA design for different CRISPR-Cas subtypes
}
\section{References}{

...
}

\section{Further details}{
 Every time CaSilico runs, it only takes one input parameter from: "fasta_file_location", "accession_number" or "seq_coordinate." It is recommended to use the "fasta_file_location" parameter as much as possible, ie to introduce the file path of your sequences to Cesilico.The method parameter takes inputs 1 or 2. In method 1, the mismatch position in the spacer is dealt with stringently, which is removed in method 2. in other words The first method is more conservative and may result in fewer crRNAs than the second method.
...
}

\examples{

# Run with fasta_file_location

data<-paste0(system.file(package = "CaSilico"),"/data/3D.fasta")
CaSilico(Results_folder_name="Example",
         fasta_file_location=data,
         Accession_Number=NULL,
         CRISPR_Types=c("casVI_A"),
         method = 1,
         mismatch_threshold=0.98,
         off_Target = F,
         ask_question = F)




# Run with fasta_file_location and off-target prediction



data<-paste0(system.file(package = "CaSilico"),"/data/3D.fasta")
genome_dir<-paste0(system.file(package = "CaSilico"),"/dependency/genomes_off_target")
location_of_off_target_sequene_file=file.path(genome_dir,list.files(genome_dir))
name_of_off_target_sequence_file=c("Genome1","Genome2","Genome3")

CaSilico(Results_folder_name="Example2",
         fasta_file_location=data,
         Accession_Number=NULL,
         CRISPR_Types=c("casVI_A","casVI_B","casVI_D","casV_A","casV_B","casV_F1"),
         method = 1,
         mismatch_threshold=0.98,
         off_Target = T,
         ask_question = F,
         Organism=NULL,
         local_file_for_off_target=T,
         location_of_off_target_sequene_file=location_of_off_target_sequene_file,
         name_of_off_target_sequence_file=name_of_off_target_sequence_file)




 # Run with fasta_file_location

 CaSilico(Results_folder_name="Example3",
          Accession_Number=c("U15717", "U15718"),
          CRISPR_Types=c("casVI_A","casVI_B"),
          method = 1,
          mismatch_threshold=0.98,
          off_Target = F,
          ask_question = F)




 # Run with coordinate
 CaSilico(Results_folder_name="Example4",
          seq_cordinate=list(chromosome=c("1","1"),
                             start=c("2000","2000"),
                             end=c("2150","2150"),
                             strand=c("-","-"),
                             species = c("bos_taurus","bos_taurus")),
          CRISPR_Types=c("casVI_B","casVI_A"),
          method = 1,
          mismatch_threshold=0.98,
          off_Target = F,
          ask_question = F)


}
