% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas_Type.R
\name{compliment}
\alias{compliment}
\title{compute and return compliment of a sequence}
\usage{
compliment(seq)
}
\arguments{
\item{seq}{sequence of DNA or RNA}
}
\value{
compliment of sequence .
}
\description{
compute and return compliment of a sequence
}
\section{References}{

...
}

\section{Further details}{

...
}

