% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam.R
\docType{data}
\name{example_seq}
\alias{example_seq}
\title{Festa file related to the Foot-and-mouth disease virus 3D gene}
\format{
An object of class `"fasta"`.
}
\source{
Foot-and-mouth disease virus, <https://www.ncbi.nlm.nih.gov/>
}
\usage{
data(example_seq)
}
\description{
Festa file related to the Foot-and-mouth disease virus 3D gene
}
\examples{
data(example_seq)
seq <- example

}
\keyword{example}
