% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas_Type.R
\name{mafftR}
\alias{mafftR}
\title{mafft (multiple sequence alignment) in R}
\usage{
mafftR(fasta_file_location, mafft_Options = "--auto   --inputorder")
}
\arguments{
\item{fasta_file_location}{path and name of fasta file for ex c:/Desktop/Example.fasta}

\item{mafft_Options}{Options for run mafft for ex  "--auto   --inputorder"}
}
\value{
msa file (output of mafft),always output format is Clustal and name is final_fasta_file.aln
}
\description{
mafft (multiple sequence alignment) in R
}
\section{Further details}{
mafftR reports output as a cluster in final_fasta_file.aln
details for Options at https://mafft.cbrc.jp/alignment/software/manual/manual.html
}

