% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas_Type.R
\name{Rcompliment}
\alias{Rcompliment}
\title{compute and return Revers compliment of a sequence}
\usage{
Rcompliment(seq)
}
\arguments{
\item{seq}{sequence of DNA or RNA}
}
\value{
Reverse compliment of sequence .
}
\description{
compute and return Revers compliment of a sequence
}
\section{References}{

...
}

\section{Further details}{

...
}

