% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas_Type.R
\name{VDotplot}
\alias{VDotplot}
\title{Dot plot for small sequence}
\usage{
VDotplot(
  sequence1,
  sequence2,
  wsize = 1,
  wstep = 1,
  nmatch = 1,
  table_col = "goldenrod1",
  points_col = "red",
  xlab = "sequence1",
  ylab = "sequence2"
)
}
\arguments{
\item{sequence1}{First sequence}

\item{sequence2}{Second sequence}

\item{wsize}{the size in chars of the moving window.}

\item{wstep}{the size in chars for the steps of the moving window. Use wstep == wsize for non-overlapping windows.}

\item{nmatch}{if the number of match per window is greater than or equal to nmatch then a dot is produced.}

\item{table_col}{table color by default table_col = "goldenrod1"}

\item{points_col}{point color by default points_col = "red"}

\item{xlab}{Label for x axis}

\item{ylab}{Label for y axis}
}
\value{
plots.
}
\description{
Dot plot for small sequence
}
\section{References}{

...
}

\section{Further details}{
In order to visualize this function better, short sequences are recommended.
...
}

