% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas_Type.R
\name{compliment}
\alias{compliment}
\title{Compute and return compliment of a sequence.}
\usage{
compliment(seq)
}
\arguments{
\item{seq}{Sequence of DNA or RNA}
}
\value{
Compliment of sequence .
}
\description{
Compute and return compliment of a sequence.
}
\section{References}{

...
}

\section{Further details}{

...
}

